toleranced moved into jitx
